# Istio config list: loop call validations per namespace in backend