--- a/tests/integration/assets/bookinfo-destination-rule-multimatch.yaml
+++ b/tests/integration/assets/bookinfo-destination-rule-multimatch.yaml
@@ -13,7 +13,7 @@ spec:
       tls:
         mode: SIMPLE
   exportTo:
-  - 'istio-system'
+  - '.'
 ---
 apiVersion: networking.istio.io/v1
 kind: DestinationRule
@@ -30,4 +30,4 @@ spec:
       tls:
         mode: SIMPLE
   exportTo:
-  - 'istio-system'
+  - '.'