### Describe the change

We had a regression at some point when calling validations per all namespaces, this caused an issue that Istio config objects are not filtered by `exportTo` field, because there is not destination namespace to check whether particular Istio config is exported to that namespace or not.

The solution here is to return back the loop per namespace to call validations.

There is another issue to be fixed later https://github.com/kiali/kiali/issues/7690
Istio config validations needs to be refactored and optimized, but this is out of scope of this PR.
This PR only returns back this loop of code missing of which was caused this validation inconsistency: https://github.com/kiali/kiali/pull/7168/files#diff-ce9a87f33646f1c782b605531625267299fc515a5754ecf697ef6e11c011ceaeL88-L99

The performance degradation for Istio Config list is 2.5 - 3.5 times.
Before:
All Namespaces Istio Configs load time: 1.0479 sec, baseline: 2 sec, difference: -0.95210 sec
Selected Namespaces Istio Configs load time: 0.49130 sec, baseline: 2 sec, difference: -1.5087 sec
After:
All Namespaces Istio Configs load time: 2.4857 sec, baseline: 2 sec, difference: +0.48570 sec
Selected Namespaces Istio Configs load time: 1.8972 sec, baseline: 2 sec, difference: -0.10280 sec

But still is ~ within 2s of baselines.

Overview page performance comparison before and after:
Before:
Init page load time: 1.4626 sec, baseline: 2 sec, difference: -0.53738 sec
After:
Init page load time: 2.4374 sec, baseline: 2 sec, difference: +0.43742 sec

All tests were run for 50 namespaces, 2 configs per namespace.
### Steps to test the PR

Steps from https://github.com/kiali/kiali/issues/7685

### Automation testing

Integration test created for this PR.

### Issue reference
https://github.com/kiali/kiali/issues/7685