@hhovsepy 2024-09-06
Added the similar loop for Overview page's namespace cards validation to avoid inconsistency. @nrfox @jshaughn ready for re-review.