@jshaughn 2024-09-10
> If I understand this correctly, this PR is just a "quick fix" to revert back to per-namespace processing, even when asking to validate all namespaces (i.e. handler is passed namespace == ""). The handler is now forcing an explicit namespace to be sent to GetValidations, bypassing the current support for namespace == "" in that function. It seems like we should still be able to fetch all namespaces and then apply exportTo filtering, as needed at the business level. Is this what we may do in a follow-up PR (maybe for 2.0?). I'm conditionally approving this as long as my understanding is correct, and we have an issue created for a near-term "real fix".

@hhovsepy Before re-review, can you respond to this prior comment? I'm trying to understand whether there is follow-on work, or whether this is the final fix.