@hhovsepy 2024-09-10
> > If I understand this correctly, this PR is just a "quick fix" to revert back to per-namespace processing, even when asking to validate all namespaces (i.e. handler is passed namespace == ""). The handler is now forcing an explicit namespace to be sent to GetValidations, bypassing the current support for namespace == "" in that function. It seems like we should still be able to fetch all namespaces and then apply exportTo filtering, as needed at the business level. Is this what we may do in a follow-up PR (maybe for 2.0?). I'm conditionally approving this as long as my understanding is correct, and we have an issue created for a near-term "real fix".
> 
> @hhovsepy Before re-review, can you respond to this prior comment? I'm trying to understand whether there is follow-on work, or whether this is the final fix.

@jshaughn oh sorry, yes right, this is a quick fix of inconsistency, which in fact is reverting the call with namespace="" for performance improvements which causes inconsistency in corner cases when object is "exportedTo" another namespace. The necessity of this fix is that Validation are designed in the way, that they prepare Istio config list in the beginning then pass this list to loop validations, this is done fore performance reasons. And the weakness of that approach is that "exportTo" filtering is also done in preparation phase, which means that without "namespace" param given it cannot filter "exportTo", because it does not know against which "namespace" to filter "exportTo". 
The better solution is eigther complete rework all validations and add "exportTo" filtering for each of them, but this is risky and requires more efforts before 2.0. Or to cache Validation results which will resolve performance issues in loop calls per  namespace.