@jshaughn 2024-09-10
> IMO before we do any major reworking of validations we need to do: https://github.com/kiali/kiali/issues/4713. Users do not want two different validation tools that potentially differ from one another.

I guess another option would be to somehow pull the `analyze` results into Kiali.  Not sure if that's feasible.  I'm ambivalent about migrating our stuff to Istio Analyze.

> Just to note that caching the validations won't actually speed up the time it takes to validate. It will only improve the user's perception of how long it takes to validate since we'll move validations outside of the request's life cycle. With caching, if validations take a long time to perform then users will likely see stale data and the Kiali server will consume more resources in performing the validations.

Right, I get that.  But I think config changes rarely.  I think I'd be fine , actually, performing validations at startup , caching it, and never doing it again unless someone either  makes a config change via Kiali, or possible requests Kiali to re-validate all config.