@nrfox 2024-09-10
> I guess another option would be to somehow pull the analyze results into Kiali. Not sure if that's feasible. I'm ambivalent about migrating our stuff to Istio Analyze.

I think this is gated behind a feature flag in istio but the results can get written to the `status` field of every istio object.